requirements.txt
urllib3
testrail_api
pytest
allure-pytest == 2.8.18